Add a console script
