+++ setup.cfg
@@ -44,6 +44,10 @@ setup_requires =
 tests_require = 
     pytest
 
+[options.entry_points]
+console_scripts =
+    demoji = demoji.main:main
+
 [options.extras_require]
 ujson = ujson
